+++ lib/barking_iguana/compound/host.rb
@@ -1,20 +1,25 @@
 module BarkingIguana
   module Compound
     class Host
-      attr_accessor :name, :uri, :state
+      attr_accessor :inventory_name, :ip_address, :state
 
-      def initialize(name:, uri:)
-        self.name = name.gsub(/[^a-z0-9\-]/, '-')
-        self.uri = uri
+      def initialize(name:, ip_address:)
+        self.inventory_name = name
+        self.ip_address = ip_address
         self.state = 'unknown'
       end
 
-      def assign_ip_address ip_address
-        self.uri = ip_address
+      def name
+        @name ||= inventory_name.gsub(/[^a-z0-9\-]/, '-')
       end
 
-      def ip_address
-        uri
+      def <=> other
+        name <=> other.name
+      end
+      include Comparable
+
+      def assign_ip_address new_ip_address
+        self.ip_address = new_ip_address
       end
 
       def ssh_key
@@ -23,6 +28,10 @@ module BarkingIguana
 
       def ssh_username
         'vagrant'
+      end
+
+      def ssh_extra_args
+        '-o StrictHostKeyChecking=no'
       end
     end
   end